backend/requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==4.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
gunicorn==25.3.0
packaging==26.0
Pillow==9.5.0
psycopg2-binary==2.9.11
python-dotenv==1.0.0
pytz==2026.1.post1
sqlparse==0.5.5
tzdata==2026.1
whitenoise==6.12.0
